+++ .eslintrc.cjs
@@ -5,16 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 module.exports = {
-  extends: [
-    'eslint-config-salesforce-typescript',
-    'eslint-config-salesforce-license',
-    'plugin:sf-plugin/recommended',
-    'xo-react/space',
-  ],
+  extends: ['eslint-config-salesforce-typescript', 'eslint-config-salesforce-license', 'plugin:sf-plugin/recommended'],
   rules: {
     // allow deleting object properties via rest operator
     '@typescript-eslint/no-unused-vars': ['error', { ignoreRestSiblings: true }],
-    'react/jsx-tag-spacing': 'off',
   },
   ignorePatterns: ['test/nuts/specialTypes/*Project/**', 'test/nuts/retrieve/partialBundleDeleteProject/**'],
 };