+++ test/nuts/convert/decompose.nut.ts
@@ -7,11 +7,17 @@
 
 import path from 'node:path';
 import fs from 'node:fs';
-import { expect } from 'chai';
-import { TestSession, execCmd } from '@salesforce/cli-plugins-testkit';
+import { expect, config } from 'chai';
+import { Interaction, TestSession, execCmd, execInteractiveCmd } from '@salesforce/cli-plugins-testkit';
 import { type ProjectJson } from '@salesforce/schemas';
+import { Messages } from '@salesforce/core/messages';
 import { SourceBehaviorResult } from '../../../src/commands/project/convert/source-behavior.js';
 import { DRY_RUN_DIR, PRESETS_PROP } from '../../../src/utils/convertBehavior.js';
+
+Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
+const messages = Messages.loadMessages('@salesforce/plugin-deploy-retrieve', 'convert.source-behavior');
+
+config.truncateThreshold = 0;
 
 describe('source behavior changes', () => {
   let session: TestSession;
@@ -54,7 +60,7 @@ describe('source behavior changes', () => {
     await fs.promises.rm(path.join(session.project.dir, DRY_RUN_DIR), { recursive: true });
   });
 
-  it('throws on a packageDir not using main/default', async () => {
+  it('warns on a packageDir not using main/default', async () => {
     const newDir = path.join(session.project.dir, 'other-dir');
     // create the new packageDir
     await fs.promises.mkdir(path.join(newDir, 'labels'), { recursive: true });
@@ -74,29 +80,21 @@ describe('source behavior changes', () => {
       )
     );
 
-    const result = execCmd('project convert source-behavior --behavior decomposeCustomLabelsBeta --json', {
-      ensureExitCode: 1,
-    });
-    expect(result.jsonOutput?.name).to.equal('PackageDirectoryNeedsMainDefaultError');
-    // put stuff back the way it was
-    await fs.promises.rm(newDir, { recursive: true });
-    await fs.promises.writeFile(
-      path.join(session.project.dir, 'sfdx-project.json'),
-      JSON.stringify(originalProject, null, 2)
-    );
-  });
-
-  it.skip('produces actual output and makes expected changes', async () => {
-    const result = execCmd<SourceBehaviorResult>(
-      'project convert source-behavior --behavior decomposeCustomLabelsBeta --json',
+    const result = await execInteractiveCmd(
+      'project convert source-behavior --behavior decomposeCustomLabelsBeta',
+      { Proceed: ['y', Interaction.ENTER] },
       {
         ensureExitCode: 0,
       }
     );
-    expect(result.jsonOutput?.result.deletedFiles).to.deep.equal([
-      path.join(session.project.dir, 'force-app', 'main', 'default', 'labels', 'CustomLabels.labels-meta.xml'),
-    ]);
-    expect(result.jsonOutput?.result.createdFiles).to.have.length(4);
+    expect(result.stderr).to.include(messages.getMessage('basicConfirmation'));
+    expect(result.stderr).to.include(messages.getMessage('mainDefaultConfirmation'));
+
+    expect(result.stdout).to.include('Deleted Files');
+    expect(result.stdout).to.include('Created Files');
+    expect(result.stdout).to.include(
+      path.join(session.project.dir, 'force-app', 'main', 'default', 'labels', 'CustomLabels.labels-meta.xml')
+    );
     // it modified the project json
     expect((await getProject(session))[PRESETS_PROP]).to.deep.equal(['decomposeCustomLabelsBeta']);
 
@@ -104,14 +102,15 @@ describe('source behavior changes', () => {
     expect(fs.existsSync(path.join(session.project.dir, DRY_RUN_DIR))).to.be.false;
   });
 
-  it.skip("throws on repeated preset that's already done", () => {
-    const err = execCmd<SourceBehaviorResult>(
+  it("throws on repeated preset that's already done", async () => {
+    const err = await execInteractiveCmd(
       'project convert source-behavior --behavior decomposeCustomLabelsBeta --json',
+      {},
       {
         ensureExitCode: 1,
       }
     );
-    expect(err.jsonOutput?.name).to.equal('sourceBehaviorOptionAlreadyExists');
+    expect(err.stdout).to.include('sourceBehaviorOptionAlreadyExists');
   });
 
   after(async () => {